Move collection related classes to collection package
