+++ chat/src/main/java/bisq/chat/channel/PrivateChannelService.java
@@ -20,7 +20,7 @@ package bisq.chat.channel;
 import bisq.chat.message.MessageType;
 import bisq.chat.message.PrivateChatMessage;
 import bisq.chat.message.Quotation;
-import bisq.common.observable.ObservableArray;
+import bisq.common.observable.collection.ObservableArray;
 import bisq.common.util.StringUtils;
 import bisq.network.NetworkService;
 import bisq.network.p2p.message.NetworkMessage;
