+++ chat/src/main/java/bisq/chat/trade/pub/PublicTradeChannelService.java
@@ -23,8 +23,8 @@ import bisq.chat.message.Quotation;
 import bisq.common.currency.Market;
 import bisq.common.currency.MarketRepository;
 import bisq.common.observable.Observable;
-import bisq.common.observable.ObservableArray;
-import bisq.common.observable.ObservableSet;
+import bisq.common.observable.collection.ObservableArray;
+import bisq.common.observable.collection.ObservableSet;
 import bisq.network.NetworkService;
 import bisq.network.p2p.services.data.storage.DistributedData;
 import bisq.network.p2p.services.data.storage.auth.AuthenticatedData;
