+++ chat/src/main/java/bisq/chat/trade/pub/PublicTradeChannelStore.java
@@ -17,8 +17,8 @@
 
 package bisq.chat.trade.pub;
 
-import bisq.common.observable.ObservableArray;
-import bisq.common.observable.ObservableSet;
+import bisq.common.observable.collection.ObservableArray;
+import bisq.common.observable.collection.ObservableSet;
 import bisq.common.proto.ProtoResolver;
 import bisq.common.proto.UnresolvableProtobufMessageException;
 import bisq.persistence.PersistableStore;
