+++ common/src/main/java/bisq/common/observable/collection/CollectionChangeMapper.java
@@ -15,7 +15,7 @@
  * along with Bisq. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package bisq.common.observable;
+package bisq.common.observable.collection;
 
 import lombok.EqualsAndHashCode;
 import lombok.ToString;
