+++ common/src/main/java/bisq/common/observable/collection/ObservableArray.java
@@ -15,7 +15,7 @@
  * along with Bisq. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package bisq.common.observable;
+package bisq.common.observable.collection;
 
 import org.jetbrains.annotations.NotNull;
