+++ common/src/main/java/bisq/common/observable/collection/ObservableCollection.java
@@ -15,8 +15,9 @@
  * along with Bisq. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package bisq.common.observable;
+package bisq.common.observable.collection;
 
+import bisq.common.observable.Pin;
 import org.jetbrains.annotations.NotNull;
 
 import java.util.Collection;
