+++ offer/src/main/java/bisq/offer/OfferBookService.java
@@ -17,7 +17,7 @@
 
 package bisq.offer;
 
-import bisq.common.observable.ObservableSet;
+import bisq.common.observable.collection.ObservableSet;
 import bisq.network.NetworkService;
 import bisq.network.p2p.services.data.DataService;
 import bisq.network.p2p.services.data.storage.auth.AuthenticatedData;
