+++ offer/src/main/java/bisq/offer/OpenOfferService.java
@@ -23,7 +23,7 @@ import bisq.account.settlement.SettlementMethod;
 import bisq.common.currency.Market;
 import bisq.common.monetary.Monetary;
 import bisq.common.monetary.Quote;
-import bisq.common.observable.ObservableSet;
+import bisq.common.observable.collection.ObservableSet;
 import bisq.common.threading.ExecutorFactory;
 import bisq.common.util.CompletableFutureUtils;
 import bisq.common.util.StringUtils;
