+++ user/src/main/java/bisq/user/identity/UserIdentityStore.java
@@ -18,7 +18,7 @@
 package bisq.user.identity;
 
 import bisq.common.observable.Observable;
-import bisq.common.observable.ObservableSet;
+import bisq.common.observable.collection.ObservableSet;
 import bisq.common.proto.ProtoResolver;
 import bisq.common.proto.UnresolvableProtobufMessageException;
 import bisq.persistence.PersistableStore;
