+++ user/src/main/java/bisq/user/role/RoleRegistrationServiceStore.java
@@ -17,7 +17,7 @@
 
 package bisq.user.role;
 
-import bisq.common.observable.ObservableSet;
+import bisq.common.observable.collection.ObservableSet;
 import bisq.common.proto.ProtoResolver;
 import bisq.common.proto.UnresolvableProtobufMessageException;
 import bisq.persistence.PersistableStore;
