+++ wallets/bitcoind/src/main/java/bisq/wallets/bitcoind/BitcoinWalletStore.java
@@ -17,7 +17,7 @@
 
 package bisq.wallets.bitcoind;
 
-import bisq.common.observable.ObservableSet;
+import bisq.common.observable.collection.ObservableSet;
 import bisq.common.proto.ProtoResolver;
 import bisq.common.proto.UnresolvableProtobufMessageException;
 import bisq.persistence.PersistableStore;
