+++ wallets/elementsd/src/main/java/bisq/wallets/elementsd/LiquidWalletStore.java
@@ -17,7 +17,7 @@
 
 package bisq.wallets.elementsd;
 
-import bisq.common.observable.ObservableArray;
+import bisq.common.observable.collection.ObservableArray;
 import bisq.common.proto.ProtoResolver;
 import bisq.common.proto.UnresolvableProtobufMessageException;
 import bisq.persistence.PersistableStore;
